+++ client/src/app/modules/pixlisecore/models/expression-data-source.ts
@@ -450,20 +450,27 @@ export class ExpressionDataSource
 
           if (idx < 0) {
             throw new Error(
-              `The currently loaded quantification does not contain column: "${dataLabel}". Please select (or create) a quantification with the relevant element.`
+              `Scan ${this._scanId} quantification does not contain column: "${dataLabel}". Please select (or create) a quantification with the relevant element.`
             );
           }
 
           //console.log('getQuantifiedDataForDetector detector='+detectorId+', dataLabel='+dataLabel+', idx='+idx+', factor='+toElemConvert);
 
-          const data = ExpressionDataSource.getQuantifiedDataValues(quantData.data, detectorId, idx, toElemConvert, dataLabel.endsWith("_%"));
+          const data = ExpressionDataSource.getQuantifiedDataValues(this._scanId, quantData.data, detectorId, idx, toElemConvert, dataLabel.endsWith("_%"));
           return PMCDataValues.makeWithValues(data);
         })
       )
     );
   }
 
-  public static getQuantifiedDataValues(quantData: Quantification, detectorId: string, colIdx: number, mult: number | null, isPctColumn: boolean): PMCDataValue[] {
+  public static getQuantifiedDataValues(
+    scanId: string,
+    quantData: Quantification,
+    detectorId: string,
+    colIdx: number,
+    mult: number | null,
+    isPctColumn: boolean
+  ): PMCDataValue[] {
     const resultData: PMCDataValue[] = [];
     let detectorFound = false;
 
@@ -510,11 +517,7 @@ export class ExpressionDataSource
       }
 
       throw new Error(
-        'The currently loaded quantification does not contain data for detector: "' +
-          detectorId +
-          '". It only contains detector(s): "' +
-          Array.from(detectors).join(",") +
-          '"'
+        `Scan ${scanId} quantification does not contain data for detector: "${detectorId}". It only contains detector(s): "${Array.from(detectors).join(",")}"`
       );
     }
 
@@ -609,7 +612,7 @@ export class ExpressionDataSource
           }
           const elemIdx = pseudoData.intensityLabels.indexOf(name);
           if (elemIdx == -1) {
-            throw new Error(`The currently loaded dataset does not include pseudo-intensity data with column name: "${name}"`);
+            throw new Error(`Scan ${this._scanId} does not include pseudo-intensity data with column name: "${name}"`);
           }
 
           // Run through all locations & build it
@@ -849,7 +852,7 @@ export class ExpressionDataSource
 
   async getRoughnessData(): Promise<PMCDataValues> {
     if (this._debug) {
-      this.logFunc(`getHousekeepingData()`);
+      this.logFunc(`getRoughnessData()`);
     }
     return await lastValueFrom(
       combineLatest([this.getDetectedDiffraction(), this.getDiffractionPeakManualList()]).pipe(
@@ -909,7 +912,7 @@ export class ExpressionDataSource
           // If it exists as a metaLabel and has a type we can return, do it
           const metaIdx = scanMetaLabelsAndTypes.metaLabels.indexOf(name);
           if (metaIdx < 0) {
-            throw new Error('The currently loaded dataset does not include housekeeping data with column name: "' + name + '"');
+            throw new Error(`Scan ${this._scanId} does not include housekeeping data with column name: "${name}"`);
           }
 
           const metaType = scanMetaLabelsAndTypes.metaTypes[metaIdx];