+++ client/src/app/modules/spectrum/widgets/spectrum-chart-widget/spectrum-energy-calibration/single-scan-energy-calibration/single-scan-energy-calibration.component.ts
@@ -2,10 +2,12 @@ import { Component, Input, OnInit } from "@angular/core";
 import { ScanItem } from "src/app/generated-protos/scan";
 import { SpectrumEnergyCalibration } from "src/app/models/BasicTypes";
 import { EnergyCalibrationService } from "src/app/modules/pixlisecore/services/energy-calibration.service";
+import { httpErrorToString } from "src/app/utils/utils";
 
 export class SingleScanEnergyCalibration {
   constructor(
     public scan: ScanItem,
+    public quantId: string,
     public calibration: SpectrumEnergyCalibration[]
   ) {}
 }
@@ -109,10 +111,29 @@ export class SingleScanEnergyCalibrationComponent implements OnInit {
   }
 
   get hasQuantCalibration(): boolean {
-    return false;
+    return (this.calibration?.quantId.length || 0) > 0;
   }
 
-  onLoadQuant() {}
+  onLoadQuant() {
+    const scanId = this.calibration?.scan.id || "";
+    if (scanId.length <= 0) {
+      alert("Failed to get scan id for quantification when loading calibration");
+    }
+
+    const quantId = this.calibration?.quantId || "";
+    if (quantId.length <= 0) {
+      alert(`Failed to get quant id set for scan ${scanId} when loading calibration`);
+    }
+
+    this._energyCalibrationService.getQuantCalibration(scanId, quantId).subscribe({
+      next: (calibration: SpectrumEnergyCalibration[]) => {
+        this.setCalibration(calibration);
+      },
+      error: (err: any) => {
+        alert(httpErrorToString(err, "Failed to get quant data for calibration"));
+      },
+    });
+  }
 
   onValueChanged(): void {
     this.readUserEntry();